+++ src/gamspy/__init__.py
@@ -23,7 +23,7 @@ from gamspy._backend.neos import NeosClient
 from gamspy._config import _set_default_options, get_option, set_options
 from gamspy._container import Container
 from gamspy._model import Model, ModelStatus, Problem, Sense, SolveStatus
-from gamspy._options import ModelInstanceOptions, Options
+from gamspy._options import FreezeOptions, ModelInstanceOptions, Options
 from gamspy._serialization import deserialize, serialize
 from gamspy._symbols import (
     Alias,
@@ -67,6 +67,7 @@ __all__ = [
     "Ord",
     "Card",
     "Options",
+    "FreezeOptions",
     "ModelInstanceOptions",
     "Expression",
     "EngineClient",