+++ CHANGELOG.md
@@ -9,6 +9,7 @@ GAMSPy 1.9.0
   - Add description argument for model objects.
   - Make GAMSPy symbols non-iterable.
   - Inherit output argument from the container in solve function if the output argument is not specified.
+  - Rename ModelInstanceOptions to FreezeOptions and add a warning for the usage of ModelInstanceOptions.
 
 GAMSPy 1.8.0
 ------------